requirements.txt
paho_mqtt==2.1.0
psutil==6.0.0
